requirements.txt
Django==4.2.11
spacy==3.8.2
numpy==1.26.4
tensorflow-cpu==2.17.0
joblib==1.4.2
scipy==1.13.1
scikit-learn==1.3.2
gunicorn==23.0.0
